+++ docs/CHANGELOG.rst
@@ -15,6 +15,7 @@ Added
 - Add a custom database connector to optimize queries and enable them to use
   database indexes
 - Add database indexes to improve search performance
+- Add database fields and triggers for full-text search in Postgres
 
 
 ===================